requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
websockets
google-genai